PlayerLocal Split - building lib-tempo
