+++ src/lib-tempo/CMakeLists.txt
@@ -11,6 +11,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridPosition.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentID.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentPlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentTransform.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreation.hpp"
@@ -18,7 +19,6 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/RenderHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/LevelRenderer.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemote.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemoteServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/TileMask.hpp"
@@ -30,6 +30,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemID.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemLevelManager.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemPlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemTransform.hpp"
 )
@@ -44,6 +45,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentGridPosition.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentID.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentPlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentTransform.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreation.cpp"
@@ -51,7 +53,6 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/RenderHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/LevelRenderer.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemote.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemoteServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/TileMask.cpp"
@@ -63,6 +64,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemID.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemLevelManager.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemPlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemTransform.cpp"
 )
