+++ src/lib-tempo/src/entity/ComponentPlayerLocal.cpp
@@ -1,0 +1,7 @@
+#include <tempo/entity/ComponentPlayerLocal.hpp>
+
+namespace tempo{
+	
+	//Currently Empty
+
+}
